x86, ptrace: Ifdef HW_BREAKPOINTS code in ptrace
Add ifdefs on CONFIG_HW_BREAKPOINTS to the hardware break point code
in x86 ptrace.c. This prepares this file for being able to disable
HW_BREAKPOINTS.

The only debug register that needs to be handled without break points
is DR6, so do that in a simple separate function without breakpoints.

Cc: fweisbec@gmail.com
Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>