intel: Remove the standard ktype nesting
Fixes [YOCTO 6710]

intel-common-standard.scc included the standard ktype, but the
preempt-rt scc files included intel-common-standard and the preempt-rt
ktype. This resulted in a double include of the standard ktype, which
caused the meta/scripts/updateme to change the source tree from
standard/preempt-rt/base to standard/standard/preempt-rt and drop
all the preempt-rt commits.

To address this, avoid the nested inclusion of ktypes by explicitly
including them in each top level BSP definition. Remove the ktype from
intel-common-standard.scc and rename it intel-common-drivers.scc, which
more accurately describes its purpose anyway. Update the top the BSPs to
use the new name and explicitly include the required ktype.

Remove some obsolete comments and clean-up the whitespace in the
top-level BSP scc files a bit.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
Cc: Bruce Ashfield <bruce.ashfield@windriver.com>

--

Please apply to linux-yocto-3.14 and to 3.17 after reverting:
intel: create intel-common-preempt-rt and use it